bot/requirements_bot.txt
python-telegram-bot==21.7
httpx==0.27.2
python-dotenv==1.0.1
